requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
backoff==2.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.3
geographiclib==2.0
geopy==2.3.0
gql==3.4.0
graphene==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
graphqlclient==0.2.4
idna==3.4
ijson==2.3
importlib-metadata==6.0.0
inflect==6.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
paho-mqtt==1.6.1
promise==2.3
psycopg2-binary
aniso8601==9.0.1
asgiref==3.6.0
pycparser==2.21
pydantic==1.10.5
pyOpenSSL==23.0.0
requests>=2.31.0
requests-pkcs12==1.15
requests-toolbelt==0.10.1
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
websocket-client==1.5.1
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
pyyaml==6.0.1
retrying==1.3.3
opencage==2.3.0

